tools/requirements.txt
asks
mysql.connector
trio
trove-classifiers
